Update 22/04
+ Defesa (bloquear bolas), Medio (recuperar bolas), Avancado (finalizacao), Lateral (cruzamentos)

+++ Jogador.java
@@ -24,7 +24,7 @@ public class Jogador {
     }
 
 
-    public Jogador (String nome, int idade, float altura, float peso) {
+    public Jogador (String nome, int idade, double altura, double peso) {
         this.nome = nome;
         this.idade = idade;
         this.altura = altura;
